solution/myapi/requirements.txt
asgiref==3.2.5
coverage==5.0.3
Django==3.0.4
djangorestframework==3.11.0
Faker==0.9.1
mixer==6.1.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.2
